feat: Split Lore page into Map and Timeline pages
Splits the combined "Lore" page into two distinct pages: "Map" and "Timeline".

- Creates map.html and timeline.html.
- Splits lore.js into map.js and timeline.js to handle page-specific logic.
- Updates the main navigation across the site to use a dropdown menu for the "Lore" section, linking to the new pages.
- Deletes the original lore.html and lore.js files.
