+++ games.html
@@ -34,7 +34,13 @@
         <nav class="main-navigation">
             <a href="index.html">Home</a>
             <a href="games.html" class="active">Releases</a>
-            <a href="lore.html">Lore</a>
+            <div class="nav-item dropdown">
+                <a href="#" class="nav-link">Lore</a>
+                <div class="dropdown-menu">
+                    <a href="map.html">Map</a>
+                    <a href="timeline.html">Timeline</a>
+                </div>
+            </div>
         </nav>
     </header>
