+++ index.html
@@ -34,7 +34,13 @@
         <nav class="main-navigation">
             <a href="index.html" class="active">Home</a>
             <a href="games.html">Releases</a>
-            <a href="lore.html">Lore</a>
+            <div class="nav-item dropdown">
+                <a href="#" class="nav-link">Lore</a>
+                <div class="dropdown-menu">
+                    <a href="map.html">Map</a>
+                    <a href="timeline.html">Timeline</a>
+                </div>
+            </div>
         </nav>
     </header>
 
@@ -43,7 +49,7 @@
             <h2 class="hero-tagline">A visual journey through the Trails/Kiseki series.</h2>
             <div class="hero-actions">
                 <a href="games.html" class="btn btn-hero">View the Releases</a>
-                <a href="lore.html" class="btn btn-hero">Explore the Lore</a>
+                <a href="map.html" class="btn btn-hero">Explore the Lore</a>
             </div>
             <p class="hero-intro">A visual display of the narrative and chronology of the longest-running interconnected JRPG series. Track every title from 2004 to upcoming releases.</p>
         </div>
