+++ src/css/style.css
@@ -135,6 +135,38 @@ main {
     background-color: rgba(233, 213, 161, 0.15);
 }
 
+/* Dropdown Menu Styles */
+.nav-item.dropdown {
+    position: relative;
+}
+
+.nav-item.dropdown .dropdown-menu {
+    display: none;
+    position: absolute;
+    top: 100%;
+    left: 50%;
+    transform: translateX(-50%);
+    background-color: var(--primary-bg);
+    border: 1px solid var(--accent-gold);
+    border-radius: 0 0 6px 6px;
+    padding: 0.5rem 0;
+    margin-top: -1px; /* Align with bottom of parent */
+    min-width: 100%;
+    z-index: 1001;
+}
+
+.nav-item.dropdown:hover .dropdown-menu {
+    display: block;
+}
+
+.dropdown-menu a {
+    display: block;
+    width: 100%;
+    text-align: center;
+    margin: 0 !important;
+    border-radius: 0 !important;
+}
+
 .arc-navigation {
     text-align: center;
     margin-top: 0;
