+++ timeline.html
@@ -1,0 +1,101 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Trails/Kiseki Series Lore Timeline - Zemurian Atlas</title>
+    <meta name="description" content="Explore the rich history and in-game chronology of the Trails (Kiseki) series with a detailed timeline of events in Zemuria.">
+    <meta name="keywords" content="Trails series lore, Kiseki lore, timeline of Zemurian history, in-game chronology, Liberl, Crossbell, Erebonia, Calvard">
+
+    <!-- Open Graph / Facebook -->
+    <meta property="og:type" content="website">
+    <meta property="og:url" content="https://jun-eau.github.io/zemurian-atlas/timeline.html">
+    <meta property="og:title" content="Trails/Kiseki Series Lore Timeline - Zemurian Atlas">
+    <meta property="og:description" content="Explore the rich history and in-game chronology of the Trails (Kiseki) series with a detailed timeline of events in Zemuria.">
+    <meta property="og:image" content="https://jun-eau.github.io/zemurian-atlas/assets/social-preview.jpg">
+
+    <!-- Twitter -->
+    <meta property="twitter:card" content="summary_large_image">
+    <meta property="twitter:url" content="https://jun-eau.github.io/zemurian-atlas/timeline.html">
+    <meta property="twitter:title" content="Trails/Kiseki Series Lore Timeline - Zemurian Atlas">
+    <meta property="twitter:description" content="Explore the rich history and in-game chronology of the Trails (Kiseki) series with a detailed timeline of events in Zemuria.">
+    <meta property="twitter:image" content="https://jun-eau.github.io/zemurian-atlas/assets/social-preview.jpg">
+    <link rel="stylesheet" href="src/css/style.css">
+    <link rel="preconnect" href="https://fonts.googleapis.com">
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
+    <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;700&family=Playfair+Display:ital,wght@1,700&display=swap" rel="stylesheet">
+    <link rel="icon" href="assets/favicon.png" type="image/png">
+    <script src="src/js/main.js" type="module" defer></script>
+</head>
+<body>
+
+    <header>
+        <h1>Zemurian Atlas</h1>
+        <nav class="main-navigation">
+            <a href="index.html">Home</a>
+            <a href="games.html">Releases</a>
+            <div class="nav-item dropdown">
+                <a href="#" class="nav-link active">Lore</a>
+                <div class="dropdown-menu">
+                    <a href="map.html">Map</a>
+                    <a href="timeline.html">Timeline</a>
+                </div>
+            </div>
+        </nav>
+    </header>
+
+    <main class="main-lore">
+        <div id="timeline-view" class="tab-content active show">
+            <div class="timeline-section">
+                <h2>Lore Timeline</h2>
+
+                <p style="font-size: 0.9em; font-style: italic; text-align: center; margin-bottom: 15px;">
+                    The timeline presented below is based on information sourced from in-game scripts and the <a href="https://kiseki.fandom.com/wiki/Timeline_of_Zemurian_history" target="_blank" rel="noopener noreferrer">Kiseki Wiki</a>.<br>Best viewed on desktop; small screens may require horizontal scrolling.
+                </p>
+
+                <div class="lore-timeline-scroll-wrapper">
+                    <div id="lore-timeline-main-container">
+                        <div id="time-axis-container">
+                            <!-- Time axis labels (years, months) will be populated by JavaScript -->
+                        </div>
+                        <div id="game-columns-container">
+                            <div class="timeline-arc-column" id="liberl-arc-column">
+                                <h3>Liberl Arc</h3>
+                                <div class="game-entries-area">
+                                    <!-- Liberl Arc game entries will be populated by JavaScript -->
+                                </div>
+                            </div>
+                            <div class="timeline-arc-column" id="crossbell-arc-column">
+                                <h3>Crossbell Arc</h3>
+                                <div class="game-entries-area">
+                                    <!-- Crossbell Arc game entries will be populated by JavaScript -->
+                                </div>
+                            </div>
+                            <div class="timeline-arc-column" id="erebonia-arc-column">
+                                <h3>Erebonia Arc</h3>
+                                <div class="game-entries-area">
+                                    <!-- Erebonia Arc game entries will be populated by JavaScript -->
+                                </div>
+                            </div>
+                            <div class="timeline-arc-column" id="calvard-arc-column">
+                                <h3>Calvard Arc</h3>
+                                <div class="game-entries-area">
+                                    <!-- Calvard Arc game entries will be populated by JavaScript -->
+                                </div>
+                            </div>
+                            <!-- Horizontal month lines will be overlaid/drawn by JavaScript -->
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </main>
+
+    <footer>
+        <p>
+            Unofficial fan project. All trademarks and copyrights are property of their respective owners. Inspired by <a href="https://trailsinthedatabase.com/">trailsinthedatabase.com</a>. <a href="https://github.com/jun-eau/zemurian-atlas">View source on GitHub</a>.
+        </p>
+    </footer>
+
+</body>
+</html>
